rename as 'jupyter_rtc_automerge'+impacts+cleanup
