+++ automerge/rust/Makefile
@@ -5,9 +5,9 @@ ifeq ($(UNAME),Darwin)
 	cargo rustc --release -- -C link-arg=-undefined -C link-arg=dynamic_lookup
 endif
 all:
-	make clean && python setup.py install && cd .. && RUST_BACKTRACE=full python run_glootalk.py
+	make clean && python setup.py install && cd .. && RUST_BACKTRACE=full python test_jupyter_rtc_automerge.py
 clean:
-	rm -rf build dist glootalk.egg-info target Cargo.lock
+	rm -rf build dist jupyter_rtc_automerge.egg-info target Cargo.lock
 pypi:
 	python3 setup.py sdist bdist_wheel
 	twine upload dist/*
