+++ automerge/rust/jupyter_rtc_automerge/__init__.py
@@ -1,0 +1,14 @@
+from .jupyter_rtc_automerge import *
+from .nbmodel import *
+
+notebookState = {}
+
+
+def init_nb():
+    notebookState.update(
+        {'nb': init_notebook(), 'shared': automerge.new_backend()})
+    return
+
+
+def get_state():
+    print(notebookState)
