+++ automerge/rust/setup.py
@@ -1,15 +1,17 @@
 from setuptools import setup
 from setuptools_rust import RustExtension
 
+
 setup(
-    name="glootalk",
+    name="jupyter_rtc_automerge",
     version="0.1.1",
-    packages=["glootalk"],
-    author="Anirrudh Krishnan", 
+    packages=["jupyter_rtc_automerge"],
+    author="Anirrudh Krishnan",
     author_email="akrishnan@quansight.com",
-    description='The glootalk python module.',
-    url="https://github.com/anirrudh/glootalk",
-    rust_extensions=[RustExtension("glootalk.glootalk", "Cargo.toml", debug=False)],
+    description='The Jupyter RTC Python module using Automerge backend.',
+    url="https://github.com/jupyterlab/rtc",
+    rust_extensions=[RustExtension(
+        "jupyter_rtc_automerge.jupyter_rtc_automerge", "Cargo.toml", debug=False)],
     include_package_data=True,
     zip_safe=False,
 )
