+++ automerge/test_jupyter_rtc_automerge.py
@@ -1,0 +1,12 @@
+import jupyter_rtc_automerge as jrtcam
+
+
+# serialized backend
+backend = jrtcam.automerge.new_backend()
+
+changes = jrtcam.automerge.get_changes(backend)
+
+
+print("\n\n\nbackend : \n", backend)
+
+print("\n\n\nchanges : \n", changes)
